Refactor place offers components
