+++ src/pages/LendPage/components/PlaceOfferTab/PlaceLiteOffer/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from './usePlaceOfferTab'
+export * from './usePlaceLiteOffer'
