+++ src/pages/LendPage/components/PlaceOfferTab/PlaceProOffer/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from './usePlaceOfferTab'
+export * from './usePlaceProOffer'
